refractor main activty to split into smaller abstract classes
