+++ app/src/main/java/com/hover/stax/transactions/TransactionDetailsFragment.kt
@@ -19,11 +19,9 @@ import com.hover.stax.R
 import com.hover.stax.contacts.StaxContact
 import com.hover.stax.databinding.FragmentTransactionBinding
 import com.hover.stax.home.MainActivity
-import com.hover.stax.navigation.NavigationInterface
-
+import com.hover.stax.home.NavigationInterface
 import com.hover.stax.utils.AnalyticsUtil.logAnalyticsEvent
 import com.hover.stax.utils.AnalyticsUtil.logErrorAndReportToFirebase
-
 import com.hover.stax.utils.DateUtils.humanFriendlyDateTime
 import com.hover.stax.utils.UIHelper
 import com.hover.stax.utils.Utils
@@ -227,7 +225,7 @@ class TransactionDetailsFragment : DialogFragment(), NavigationInterface {
         if (!isFullScreen) {
             binding.transactionDetailsCard.setTitle(viewModel.transaction.value?.generateLongDescription(action, viewModel.contact.value, requireContext()))
         }
-        binding.infoCard.detailsNetwork.text = action.from_institution_name
+        binding.infoCard.detailsNetwork.text = action?.from_institution_name
         updateStatus(action, viewModel.transaction.value)
     }
